# specs/openapi/MULO_OpenAPI_v2.7.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MULO API",
    "version": "2.7",
    "description": "MULO 통합 API 계약 기준 OpenAPI 문서입니다.\n\n- 현재 feature 구현과 기존 OpenAPI v2.6 계약을 함께 보존합니다.\n- 각 작업의 x-implementation-status는 implemented, contract-only, excluded-v1 중 하나입니다.\n- 홈 인기 자물쇠는 조회 시점 기준 최근 7일 데이터를 사용합니다.\n- 내 자물쇠 조회와 사용자 자물쇠 지역 그룹에는 7일 제한을 적용하지 않습니다.\n- 법정동 조회가 정상 완료됐지만 결과가 없으면 legalDongCode/legalDongName을 null로 허용합니다.\n- 커서 기반 목록 API size는 기본 20, 최소 1, 최대 100이며 채팅 메시지 내역만 기본 30입니다.\n- Refresh/CSRF Cookie의 SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict입니다.\n- 서버 URL은 아직 확정되지 않아 지정하지 않았습니다."
  },
...
  "paths": {
...
    "/api/users/signup": {
      "post": {
        "tags": [
          "USER / 사용자"
        ],
        "summary": "회원가입",
        "operationId": "postApiUsersSignup",
        "description": "필수\n    email: String\nnickname: String\n    password: String",
        "responses": {
          "201": {
            "description": "새 사용자 리소스가 생성되므로 201.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "회원가입이 완료되었습니다."
                }
              }
            }
          },
          "400": {
            "description": "필수 입력값 누락은 유효성 검증 실패이므로 400.\n\n추가 케이스: 입력 형식/길이가 정책을 만족하지 않으므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "nickname",
                      "code": "NICKNAME_REQUIRED",
                      "message": "닉네임을 입력해주세요."
                    },
                    {
                      "field": "email",
                      "code": "EMAIL_REQUIRED",
                      "message": "이메일을 입력해주세요."
                    },
                    {
                      "field": "password",
                      "code": "PASSWORD_REQUIRED",
                      "message": "비밀번호를 입력해주세요."
                    }
                  ]
                }
              }
            }
          },
          "409": {
            "description": "정상 형식의 요청이 기존 활성 사용자 자원과 중복되므로 409.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "DUPLICATE_RESOURCE",
                  "message": "이미 사용 중인 정보가 있습니다.",
                  "errors": [
                    {
                      "field": "email",
                      "code": "EMAIL_DUPLICATED",
                      "message": "이미 사용 중인 이메일입니다."
                    },
                    {
                      "field": "nickname",
                      "code": "NICKNAME_DUPLICATED",
                      "message": "이미 사용 중인 닉네임입니다."
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "nickname",
                  "password"
                ]
              },
              "example": {
                "nickname": "홍길동",
                "email": "example@muro.com",
                "password": "Test1234!"
              }
            }
          }
        },
        "x-implementation-status": "implemented"
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "USER / 사용자"
        ],
        "summary": "마이페이지 내 정보 조회",
        "operationId": "getApiUsersMe",
        "description": "외부 API에서도 내부 user_id를 userId로 사용한다.",
        "responses": {
          "200": {
            "description": "현재 인증된 현재 사용자의 회원 정보를 정상적으로 조회했으므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "nickname": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "회원 정보 조회 성공",
                  "data": {
                    "userId": 35,
                    "nickname": "뮤로",
                    "email": "example@mulo.com"
                  }
                }
              }
            }
          },
          "401": {
            "description": "현재 사용자 정보를 조회하기 위한 유효한 인증 정보가 없거나, 탈퇴 사용자여서 활성 사용자로 인증할 수 없으므로 401.\\",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": [

            ]
          }
        ],
        "x-implementation-status": "implemented"
      },
      "delete": {
        "tags": [
          "USER / 사용자"
        ],
        "summary": "회원 탈퇴",
        "operationId": "deleteApiUsersMe",
        "description": "비밀번호 확인 후 users.deleted_at을 설정하고 Refresh Token을 revoke한다. soft delete 정책.",
        "responses": {
          "200": {
            "description": "회원 탈퇴 처리가 완료되었고 별도 반환 리소스가 필요하지 않으므로 200.\n\n기존 사용자 리소스를 소프트 딜리트 방식으로 정상적으로 탈퇴 처리했으므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "회원 탈퇴가 완료되었습니다."
                }
              }
            }
          },
          "400": {
            "description": "필수값 누락은 400.\n\n회원 탈퇴에 필요한 필수 입력값인 비밀번호가 누락되어 유효성 검증에 실패했으므로 400.\n\n추가 케이스: 현재 사용자 확인을 위한 입력 검증 실패이므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "password",
                      "code": "PASSWORD_REQUIRED",
                      "message": "비밀번호를 입력해주세요."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "탈퇴 대상 사용자 식별에 인증이 필요하므로 401.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "500": {
            "description": "회원 탈퇴 처리 과정에서 MULO 백엔드 내부의 예상하지 못한 처리 실패가 발생했으므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "password": "..."
              }
            }
          }
        },
        "x-implementation-status": "contract-only",
        "x-csrf-required": true
      }
    },
    "/api/users/me/nickname": {
      "patch": {
        "tags": [
          "USER / 사용자"
        ],
        "summary": "닉네임 변경",
        "operationId": "patchApiUsersMeNickname",
        "description": "현재 닉네임과 동일한 값으로 변경을 요청하면 409로 처리한다. 다른 활성 사용자가 사용 중인 닉네임도 409.",
        "responses": {
          "200": {
            "description": "닉네임 변경이 정상 반영되었으므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "nickname": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "message": "닉네임이 변경되었습니다.",
                  "data": {
                    "nickname": "새닉네임"
                  }
                }
              }
            }
          },
          "400": {
            "description": "필수값/길이 검증 실패는 400.\n\n추가 케이스: 닉네임 형식 오류 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "nickname",
                      "code": "NICKNAME_REQUIRED",
                      "message": "닉네임을 입력해주세요."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "사용자 닉네임 변경은 인증이 필요하므로 401.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "409": {
            "description": "다른 활성 사용자의 닉네임과 충돌하므로 409.\n\n추가 케이스: 요청 값은 유효하지만 현재 사용자 상태와 충돌하므로 409.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "NICKNAME_DUPLICATED",
                  "message": "이미 사용 중인 닉네임입니다."
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "nickname": "새닉네임"
              }
            }
          }
        },
        "x-implementation-status": "implemented",
        "x-csrf-required": true
      }
    },
    "/api/users/me/password": {
      "patch": {
        "tags": [
          "USER / 사용자"
        ],
        "summary": "비밀번호 변경",
        "operationId": "patchApiUsersMePassword",
        "description": "newPasswordConfirm은 프론트에서만 검증한다. 변경 후 현재 로그인 세션은 유지한다.",
        "responses": {
          "200": {
            "description": "기존 사용자 자원의 비밀번호 변경이 성공했으므로 200.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "비밀번호가 변경되었습니다."
                }
              }
            }
          },
          "400": {
            "description": "비밀번호 변경에 필요한 필수 입력값이 누락되어 유효성 검증에 실패했으므로 400.\n\n추가 케이스: 요청 형식은 정상이나 입력한 현재 비밀번호가 실제 비밀번호와 일치하지 않아 변경 조건을 충족하지 못하므로 400.\n\n추가 케이스: 새 비밀번호가 비밀번호 형식 및 길이 정책을 만족하지 않으므로 400.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "field": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "code": "INVALID_INPUT_VALUE",
                  "message": "입력값을 확인해주세요.",
                  "errors": [
                    {
                      "field": "currentPassword",
                      "code": "CURRENT_PASSWORD_REQUIRED",
                      "message": "현재 비밀번호를 입력해주세요."
                    },
                    {
                      "field": "newPassword",
                      "code": "NEW_PASSWORD_REQUIRED",
                      "message": "새 비밀번호를 입력해주세요."
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "비밀번호 변경은 인증된 활성 사용자만 수행할 수 있으므로, 유효한 인증 정보가 없는 경우 401.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "로그인이 필요합니다."
                }
              }
            }
          },
          "409": {
            "description": "요청값 자체는 유효하지만 현재 비밀번호와 동일한 값으로 변경을 요청하여 현재 사용자 상태와 충돌하므로 409.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "SAME_PASSWORD",
                  "message": "새 비밀번호는 현재 비밀번호와 달라야 합니다."
                }
              }
            }
          },
          "500": {
            "description": "MULO 백엔드 내부의 예상하지 못한 처리 실패이므로 500.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "code": "INTERNAL_SERVER_ERROR",
                  "message": "서버 내부 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currentPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "currentPassword": "Test1234!",
                "newPassword": "New1234!"
              }
            }
          }
        },
        "x-implementation-status": "implemented",
        "x-csrf-required": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access Token 인증이 필요한 API에서 사용"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HttpOnly Refresh Token Cookie. Path=/api/auth, 운영 Secure=true. SameSite 최종 정책은 배포 구조에 따라 확정하며 현재 구현은 Strict."
      },
      "xsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "description": "GET /api/csrf가 발급한 XSRF-TOKEN Cookie의 원본 값을 상태 변경 요청에 전달한다. POST /api/users/signup은 예외다."
      }
    },
...
  }
}